% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{filters-package}
\alias{filters}
\alias{filters-package}
\title{filters Package}
\description{
Enables filtering datasets by a priori specified identifiers which correspond to saved filter expressions.
}
