% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_filter}
\alias{get_filter}
\title{Get a Filter Definition}
\usage{
get_filter(id)
}
\arguments{
\item{id}{\code{character}. The filter ID}
}
\value{
A \code{list} with elements \code{title}, \code{target} and \code{condition}
}
\description{
Get a Filter Definition
}
\examples{
get_filter("SE")
get_filter("SER")

## Filter `FOO` does not exist
try(get_filter("FOO"))

}
\author{
Thomas Neitmann (\code{neitmant})
}
