% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_filters}
\alias{get_filters}
\title{Get Multiple Filter Definitions}
\usage{
get_filters(ids)
}
\arguments{
\item{ids}{\code{character}. The filter IDs as a single string separated by underscores}
}
\value{
A named \code{list} of filter definitions
}
\description{
Get Multiple Filter Definitions
}
\examples{
get_filters("REL_SER")
get_filters("OS_IT")

}
\author{
Thomas Neitmann (\code{neitmant})
}
