% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hampel}
\alias{hampel}
\title{Hampel Filter for outlier removal}
\usage{
hampel(x, window_size, a = 3, k = 1.4826, recentre = TRUE, miss = "single")
}
\arguments{
\item{x}{Input time-series as a vector.}

\item{window_size}{Window size as an integer}

\item{a}{Threshold number of standard deviations from the median absolute deviation to 
classify outliers as.}

\item{k}{Tuning parameter to convert standard deviations to a normal CDF.}

\item{recentre}{Whether to recentre the window around the middle point in the window so as to avoid
adding lag. I.e. for an odd window_length y[i] = mean(x[(i - floor(window_length/2)) : (i + floor(window_length/2))])
For an even window length the output is located more closely to earlier points, i.e.
y[i] = mean(x[(i - window_length/2 + 1) : (i + window_length/2)])}

\item{miss}{How to handle missing values. 'single' only returns NA for y[i] if x[i] = NA.
'all' returns NA for y[i] if any value in the window is missing, and 'none' omits NAs from the window,
so will only return NA for y[i] if all windowed values are also NA.}
}
\value{
A filtered time-series the same length as x.
}
\description{
Smoothes a series by removing outliers.
Outliers are detected by being more than a standard deviations from the
median absolute deviation.
Non-outliers aren't affected.
The a parameter scales the standard deviations to have the CDF of a normal
distribution.
}
