% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{LInSARRFE}
\alias{LInSARRFE}
\title{Phase filters}
\usage{
LInSARRFE(imageRaster, param, eth, xi = 0.9)
}
\arguments{
\item{imageRaster}{Phase difference image (interferogram) with values ranging from
\code{-pi to pi}.}

\item{param}{Vector, \code{param<-c(r, theta, L)}.}

\item{eth}{Mask threshold to determine the maximum magnitude
(defined empirically)}

\item{xi}{Percentage of all pixels within the selected window.}

\item{r}{Complex correction coefficient (coherence), \code{0<r<=1}.}

\item{theta}{Is a potential mean between, \code{-pi<theta<=pi}.}

\item{L}{Number of looks or targeted, \code{L>1} integer.}
}
\value{
A filtered image
}
\description{
The Extension of the Refined Lee InSAR Filter (LInSARRFE) was
             developed based on the multilook phase difference expression
             'dFuncGierullEq7', implemented in the FDDPhase' package and
             defined in 'Closed-Form Expressions for InSAR Sample Statistics
             and Its Application to Non-Gaussian Data' for integer values
             of \code{L}, positive, and greater than \code{1}.
}
\details{
Phase filters in the spatial domain, used to reduce phase errors
         or noise.
}
\examples{
# Example of function usage
img_LInSARRFE <- LInSARRFE(imageRaster, param=c(0.7, 0, 2), eth=2, xi=0.9)

}
\author{
Araújo, J. E.
}
