% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{list_all_filters}
\alias{list_all_filters}
\title{List All Filters}
\usage{
list_all_filters()
}
\value{
A \code{data.frame} with columns \code{id}, \code{title}, \code{target} and \code{condition}
}
\description{
List all available filters
}
\examples{
list_all_filters()

}
\author{
Thomas Neitmann (\code{neitmant})
}
