% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{TcCfilter}
\alias{TcCfilter}
\title{Phase filters}
\usage{
TcCfilter(imageRaster, param, eth, xi = 0.9)
}
\arguments{
\item{imageRaster}{Phase difference image (interferogram) with values ranging from
\code{-pi to pi}.}

\item{param}{Vector, \code{param<-c(mu, sigma)}.}

\item{eth}{Mask threshold to determine the maximum magnitude
(defined empirically)}

\item{xi}{Percentage of all pixels within the selected window.}

\item{mu}{loc: Floating point tensor; the modes of the corresponding
non-truncated Cauchy distribution(s)}

\item{sigma}{scale: Floating point tensor; the scales of the distribution(s).
Must contain only positive values.}
}
\value{
A filtered image
}
\description{
The Truncated Circular Cauchy Filter (TcCfilter) was
             developed based on the empirical Truncated Circular Cauchy
             model,implemented in the 'FDDPhase (function dTruncCauchy)'
             package.
}
\details{
Phase filters in the spatial domain, used to reduce phase errors
         or noise.
}
\examples{
# Example of function usage
img_TcCfilter <- TcCfilter(imageRaster, param=c(-0.9, 0.2), eth=2, xi=0.9)

}
\author{
Araújo, J. E.
}
