% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{TcNfilter}
\alias{TcNfilter}
\title{Phase filters}
\usage{
TcNfilter(imageRaster, param, eth, xi = 0.9)
}
\arguments{
\item{imageRaster}{Phase difference image (interferogram) with values ranging from
\code{-pi to pi}.}

\item{param}{Vector, \code{param<-c(mu, sigma)}.}

\item{eth}{Mask threshold to determine the maximum magnitude
(defined empirically)}

\item{xi}{Percentage of all pixels within the selected window.}

\item{mu}{loc: Floating point tensor; the means of the distribution(s)}

\item{sigma}{scale: Loating point tensor; the stddevs of the distribution(s).
Must contain only positive values.}
}
\value{
A filtered image
}
\description{
The Truncated Circular Normal Filter (TcNfilter) was
             developed based on the empirical Truncated Circular Normal
             model,implemented in the 'FDDPhase (function dTruncNorm)' package.
}
\details{
Phase filters in the spatial domain, used to reduce phase errors
         or noise.
}
\examples{
# Example of function usage
img_TcNfilter <- TcNfilter(imageRaster, param=c(-0.3, 0.5), eth=2, xi=0.9)

}
\author{
Araújo, J. E.
}
