% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ukf.R
\name{ukf.compute}
\alias{ukf.compute}
\title{Compute the unscented Kalman Filter}
\usage{
ukf.compute(m0, P0, y, f, g, Q, R, n = length(m), alpha = 0.001,
  kappa = 0, betta = 2)
}
\arguments{
\item{m0}{The initial mean}

\item{P0}{The initial covariance}

\item{y}{The sequence of signals}

\item{f}{The law of motion for the state}

\item{g}{The measurement equation}

\item{Q}{The state shock covariance}

\item{R}{The measurement covariance}

\item{n}{The dimension of the problem}
}
\value{
The mean and covariance of the predicted state at each point in time
}
\description{
Compute the unscented Kalman Filter
}

