% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ukf.R
\name{ukf.predict}
\alias{ukf.predict}
\title{Predict the state for the unscented Kalman Filter}
\usage{
ukf.predict(m, P, f, Q, n = length(m), alpha = 0.001, kappa = 0,
  betta = 2)
}
\arguments{
\item{m}{The previous mean}

\item{P}{The previous covariance}

\item{f}{The law of motion for the state}

\item{Q}{The state covariance}

\item{n}{The dimension of the problem}
}
\value{
The mean and covariance of the predicted state
}
\description{
Predict the state for the unscented Kalman Filter
}

