% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ukf.R
\name{ukf.update}
\alias{ukf.update}
\title{Predict the state for the unscented Kalman Filter}
\usage{
ukf.update(m, P, g, R, y, n = length(m), alpha = 0.001, kappa = 0,
  betta = 2)
}
\arguments{
\item{m}{The predicted mean}

\item{P}{The predicted covariance}

\item{g}{The measurement equation}

\item{R}{The measurement covariance}

\item{n}{The dimension of the problem}
}
\value{
The mean and covariance of the predicted state
}
\description{
Predict the state for the unscented Kalman Filter
}

