% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{filters-package}
\alias{filters}
\alias{filters-package}
\title{filters Package}
\description{
Enables filtering datasets by a prior specified identifiers which correspond to saved filter expressions.
}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} [contributor]

Authors:
\itemize{
  \item Thomas Neitman (Original creator of the package)
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
